+++ src/App.vue
@@ -29,6 +29,16 @@
   body{
     margin: 0;
     padding: 0;
+
+    .icon-map-control-question {
+      transition: transform 1s;
+      transform: skew(0deg, 0deg);
+    }
+
+    .icon-map-control-question:hover {
+      transition: transform 1s;
+      transform: skew(-15deg, 10deg);
+    }
   }
   #app {
     font-family: 'Avenir', Helvetica, Arial, sans-serif;
@@ -47,6 +57,22 @@
       display: flex;
       flex-direction: column;
     }
-
   }
+  /* Styles for Custom Map Controls (the button that are within the map) */
+  .mapboxgl-ctrl .icon-map-control svg {
+    width: 18px;
+    height: 18px;
+    margin-top: 4px;
+  }
+
+  .icon-map-control-question {
+    transition: transform 1s;
+    transform: skew(0deg, 0deg);
+  }
+
+  .icon-map-control-question:hover {
+    transition: transform 1s;
+    transform: skew(-15deg, 10deg);
+  }
+
 </style>