+++ src/router.js
@@ -1,11 +1,13 @@
 import Vue from "vue";
 import Router from "vue-router";
 import MapBox from './components/MapBox.vue';
-import About from './components/About.vue';
+import QuestionsAndAnswers from './views/QuestionsAndAnswers.vue';
 import Error404 from "./views/Error404";
 
-
 Vue.use(Router);
+function lazyLoad(view){
+    return() => System.import(`@/views/${view}.vue`)
+}
 
 export default new Router({
     routes: [
@@ -15,14 +17,14 @@ export default new Router({
             component: MapBox
         },
         {
-            path: '/about',
-            name: 'About',
-            component: About
+            path: '/questionsandanswers',
+            name: 'QuestionsAndAnswers',
+            component: lazyLoad('QuestionsAndAnswers')
         },
         {
             path: "/404",
             name: "Error404",
-            component: Error404
+            component: lazyLoad('Error404')
         },
         {
             path: "*",