+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+- Added question/answers page and link on map
+- reformatted About page
 - Rivers turn on no matter the zoom
 - Changed to hash routing to accommodate the new nested folder structure on AWS
 - Added a custom 404 error page